Make restore tool easier to use; Fix docs.
The mongo_restore_database tool, with no backup files listed, would silently
not recover anything.

This changes it to search for bson files under the backup_dir, and recover those.

system.users and system.indexes are skipped, as these have different recovery
requirements, and are better re-added via Salt.

In the process, fix the documentation to reflect the above, and fix the name of the
script!